No longer specifying "Authorization" header if it is unnecessary
Previously, the "authorization" header would be present in all requests, even if the token was not set. In some circumstances this causes Kubernetes to respond with "unauthorized". Additionally, fixed how the authorization header is set after authentication to prevent a crash if no headers had prevously been set on the request